refactor(runtime): extract provider runner + hooks from supervisor (989→779)
- build_provider_runner — LLM client construction + provider match
  (ollama/anthropic/openai) moved to supervisor_runner.rs
- prepare_runtime — telemetry writer + notification routing + hook
  chain + skills loading extracted to supervisor_runner.rs
- Removed unused imports (LlmClient, HostGuard, NetworkOutboundMode,
  SystemClock, HookCtx, TelemetryEmitter, WriterTelemetryEmitter)

CLAUDE.md §4 800-line cap: 989 → 779.
